Refactor code to support additional API version
This commit moves the original API code to the v1 directory and
creates a new layer to support inclusion of a v2 API.

Also included are fixes to move tests over to use the responses
testing framework for requests and pep8
