+++ freshdesk/v1/sample_json_data/customer.json
